+++ modi/modi.py
@@ -9,7 +9,7 @@ from importlib import import_module as im
 
 from modi._exe_thrd import ExeThrd
 from modi.util.connection_util import is_network_module_connected, is_on_pi
-from modi.util.miscellaneous import ModuleList
+from modi.util.miscellaneous_util import ModuleList
 from modi.util.stranger import check_complete
 from modi.util.topology_manager import TopologyManager
 from modi.util.firmware_updater import STM32FirmwareUpdater
