+++ tests/module/input_module/test_gyro.py
@@ -2,7 +2,7 @@ import unittest
 
 from modi.module.input_module.gyro import Gyro
 from modi.util.message_util import parse_message
-from modi.util.miscellaneous import MockConn
+from modi.util.miscellaneous_util import MockConn
 
 
 class TestGyro(unittest.TestCase):
