+++ tests/module/input_module/test_ir.py
@@ -2,7 +2,7 @@ import unittest
 
 from modi.module.input_module.ir import Ir
 from modi.util.message_util import parse_message
-from modi.util.miscellaneous import MockConn
+from modi.util.miscellaneous_util import MockConn
 
 
 class TestIr(unittest.TestCase):
