+++ tests/module/input_module/test_mic.py
@@ -3,7 +3,7 @@ import unittest
 
 from modi.module.input_module.mic import Mic
 from modi.util.message_util import parse_message
-from modi.util.miscellaneous import MockConn
+from modi.util.miscellaneous_util import MockConn
 
 
 class TestMic(unittest.TestCase):
